refactor: split call and send
