+++ crates/bundle/src/call.rs
@@ -2,12 +2,10 @@
 use alloy::{
     eips::{eip2718::Encodable2718, BlockNumberOrTag},
     primitives::{keccak256, Address, Bytes, B256, U256},
-    rpc::types::mev::{EthCallBundle, EthCallBundleResponse, EthSendBundle},
+    rpc::types::mev::{EthCallBundle, EthCallBundleResponse},
 };
 use serde::{Deserialize, Serialize};
-use signet_zenith::SignedOrder;
 use std::collections::BTreeMap;
-use trevm::Block;
 
 /// Bundle of transactions for `signet_callBundle`.
 ///
@@ -225,89 +223,8 @@ impl SignetCallBundle {
     }
 }
 
-impl Block for SignetCallBundle {
-    fn fill_block_env(&self, block_env: &mut trevm::revm::primitives::BlockEnv) {
-        block_env.number =
-            self.bundle.state_block_number.as_number().map(U256::from).unwrap_or(block_env.number);
-        block_env.timestamp = self.bundle.timestamp.map(U256::from).unwrap_or(block_env.timestamp);
-        block_env.gas_limit = self.bundle.gas_limit.map(U256::from).unwrap_or(block_env.gas_limit);
-        block_env.difficulty =
-            self.bundle.difficulty.map(U256::from).unwrap_or(block_env.difficulty);
-        block_env.basefee = self.bundle.base_fee.map(U256::from).unwrap_or(block_env.basefee);
-    }
-}
-
 /// Response for `signet_callBundle`.
 pub type SignetCallBundleResponse = EthCallBundleResponse;
-
-/// Bundle of transactions for `signet_sendBundle`.
-///
-/// The Signet bundle contains the following:
-///
-/// - A standard [`EthSendBundle`] with the transactions to simulate.
-/// - A signed permit2 order to be applied on Ethereum with the bundle.
-///
-/// This is based on the flashbots `eth_sendBundle` bundle. See [their docs].
-///
-/// [their docs]: https://docs.flashbots.net/flashbots-auction/advanced/rpc-endpoint
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct SignetEthBundle {
-    /// The bundle of transactions to simulate. Same structure as a Flashbots [`EthSendBundle`] bundle.
-    #[serde(flatten)]
-    pub bundle: EthSendBundle,
-    /// Host fills to be applied with the bundle, represented as a signed
-    /// permit2 order.
-    #[serde(default)]
-    pub host_fills: Option<SignedOrder>,
-}
-
-impl SignetEthBundle {
-    /// Returns the transactions in this bundle.
-    pub fn txs(&self) -> &[Bytes] {
-        &self.bundle.txs
-    }
-
-    /// Returns the block number for this bundle.
-    pub const fn block_number(&self) -> u64 {
-        self.bundle.block_number
-    }
-
-    /// Returns the minimum timestamp for this bundle.
-    pub const fn min_timestamp(&self) -> Option<u64> {
-        self.bundle.min_timestamp
-    }
-
-    /// Returns the maximum timestamp for this bundle.
-    pub const fn max_timestamp(&self) -> Option<u64> {
-        self.bundle.max_timestamp
-    }
-
-    /// Returns the reverting tx hashes for this bundle.
-    pub fn reverting_tx_hashes(&self) -> &[B256] {
-        self.bundle.reverting_tx_hashes.as_slice()
-    }
-
-    /// Returns the replacement uuid for this bundle.
-    pub fn replacement_uuid(&self) -> Option<&str> {
-        self.bundle.replacement_uuid.as_deref()
-    }
-}
-
-/// Response for `signet_sendBundle`.
-///
-/// This is based on the flashbots `eth_sendBundle` response. See [their docs].
-///
-/// [their docs]: https://docs.flashbots.net/flashbots-auction/advanced/rpc-endpoint
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct SignetEthBundleResponse {
-    /// The bundle hash of the sent bundle.
-    ///
-    /// This is calculated as keccak256(tx_hashes) where tx_hashes are the
-    /// concatenated transaction hashes.
-    pub bundle_hash: B256,
-}
 
 #[cfg(test)]
 mod test {
@@ -316,9 +233,6 @@ mod test {
         eips::BlockNumberOrTag,
         primitives::{Address, U256},
         rpc::types::mev::{EthCallBundle, EthCallBundleTransactionResult},
-    };
-    use signet_zenith::HostOrders::{
-        Output, Permit2Batch, PermitBatchTransferFrom, TokenPermissions,
     };
 
     #[test]
@@ -351,45 +265,6 @@ mod test {
     }
 
     #[test]
-    fn send_bundle_ser_roundtrip() {
-        let bundle = SignetEthBundle {
-            bundle: EthSendBundle {
-                txs: vec![b"tx1".into(), b"tx2".into()],
-                block_number: 1,
-                min_timestamp: Some(2),
-                max_timestamp: Some(3),
-                reverting_tx_hashes: vec![B256::repeat_byte(4), B256::repeat_byte(5)],
-                replacement_uuid: Some("uuid".to_owned()),
-            },
-            host_fills: Some(SignedOrder {
-                permit: Permit2Batch {
-                    permit: PermitBatchTransferFrom {
-                        permitted: vec![TokenPermissions {
-                            token: Address::repeat_byte(66),
-                            amount: U256::from(17),
-                        }],
-                        nonce: U256::from(18),
-                        deadline: U256::from(19),
-                    },
-                    owner: Address::repeat_byte(77),
-                    signature: Bytes::from(b"abcd"),
-                },
-                outputs: vec![Output {
-                    token: Address::repeat_byte(88),
-                    amount: U256::from(20),
-                    recipient: Address::repeat_byte(99),
-                    chainId: 100,
-                }],
-            }),
-        };
-
-        let serialized = serde_json::to_string(&bundle).unwrap();
-        let deserialized: SignetEthBundle = serde_json::from_str(&serialized).unwrap();
-
-        assert_eq!(bundle, deserialized);
-    }
-
-    #[test]
     fn call_bundle_resp_ser_roundtrip() {
         let resp = EthCallBundleResponse {
             bundle_hash: B256::repeat_byte(1),
@@ -415,16 +290,6 @@ mod test {
 
         let serialized = serde_json::to_string(&resp).unwrap();
         let deserialized: SignetCallBundleResponse = serde_json::from_str(&serialized).unwrap();
-
-        assert_eq!(resp, deserialized);
-    }
-
-    #[test]
-    fn send_bundle_resp_ser_roundtrip() {
-        let resp = SignetEthBundleResponse { bundle_hash: B256::repeat_byte(1) };
-
-        let serialized = serde_json::to_string(&resp).unwrap();
-        let deserialized: SignetEthBundleResponse = serde_json::from_str(&serialized).unwrap();
 
         assert_eq!(resp, deserialized);
     }
